Make functionize various lua components, add char_defaults() component
The target skill setting, persistent message, and char dump code has
been functionized so that adding these components consists of loading
the code and adding a single line to ready() instead of blocks of code.

The target skill code, enabled now through the function target_skill(),
now as has been cleaned up in a few ways for clarity, moving some state
out of c_persist and generalizing a bit to work with char_defaults().

char_defaults() loads enabled skills on turn 0 from default settings
stored in c_persist for each unique crawl character (race + species). It
cooperates with target_skill() and loads the initial skill target as
well. It saves defaults for a char on turn 0 if it sees that no defaults
are in c_persist and that target_skill() is active. Defaults can be
manually saved through the function save_char_defaults() by e.g. binding
a macro key to the target `===save_char_defaults'.