Split lua code out of gammafunk.rc into standalone files
This makes it easier to maintain and easier for others to use individual
components. I use a script to include the lua files to make a final
version of gammafunk.rc.

This commit also removes the squishy_150hp, splits squishy_90hp into
that and squishy_60hp, and makes various adjustments to the patterns.