Add bread_swing.lua, a automatic bread swinging mode
This is an initial working version. One remaining issue is that the code
interrupts swinging with any unrecognized message. There's code to
filter out status change messages related to poison and regen, but this
approach doesn't work well for things like allies that generate messages
of speech text. Might need to look into filtering out messages by
channel, if that information is available for a given message.