Support Weapon and Ranged skill classes in char_defaults
Now setting a weapon skill (including UC) will cause that value to be
saved under a key of "Weapon" to be used for the same combination even
if a different weapon type is selected at start, based on whichever
weapon skill has base skill > 0. This is likewise done for the ranged
skills, which includes Throwing.

Documentation is updated, some bugs fix, and target_skill and
char_defaults have been made so they can fully function independently,
aside from the automatic saving/loading functionality that requires both
components to be loaded.