update-rc.py: Update rc files on WebTiles servers over WebSockets
Using a single rc file, update a set of DCSS games (versions) on a set
of WebTiles servers over WebSockets. Default locations for rc files are
`~/.crawlrc' and `~/.crawl/init.txt'.

I need to add websocket entries for all current webtiles servers. Needs
some reworking to not use global variables and to have a better way to
handle games. Probably the server argument should just be done as
required parameters as well.

Depending on how fragile the websockets/asyncio combination proves to
be, I might need to redo this using tornado. There are weird inflate
errors happening with zlib where the library thinks the data is
corrupted, depending on how the croutines as set up and schedules, and I
have no idea why.