bread_swing: Implement walk-resting
Now Nagas and Chei worshipers and sufficiently slow-moving users of
unswappable weapons can use the bread_swing functions to slow-walk-rest
back and forth in a safe, automatically determined location. It keeps
the player in or out of water based on whether we were in or out of
water initially and operates with the same restrictions as bread
swinging in terms of status changes, new event messages, and hostile
monsters. Needs documentation.

This commit also improves the message detection, ignoring messages
generated from RandomTiles and adding an `ignore_message` variable to
configure additional ones. We automatically remove preceding and
trailing control characters and newlines from any messages involved.