toggle_options.lua: Generalize to allow arbitrary option groups
This now looks for a table called option_groups that gives groups of
options to enable using a template system for the on and off states. It
creates toggle functions for each group defined in this table, which set
the options using crawl.setopt(). Also add some basic documentation to
the README.