RandomTiles: Add some missing monster and alphabetize
I'm going to add basically every monster eventually. Hopefully we can
get crawl to make a list of types of monsters available to clua at some
point. That together with merging my weapon and shield offsets into the
in-game tile offsets would make most of these entries not necessary, and
the only ones remaining would be those with special randomization or
tile usage.