Reorganize spell letters
Use a simpler system where self-enchants/escape spells use vowels+s&r,
other spells use consonants, and a system that favors using letters at
the beggining, end, or middle of words in that order of preference,
with the first letter of the school as a fallback. Trying to match
"fire" spells with fire and the like was generally a pain and the
association wasn't terribly helpful.