gammafunk.rc: Fix item/menu colors for books
Manuals were previously getting brown color even when not
identified. The prefixes made for manuals in crawl are not quite
correct, as it turns out, and the added entries work around this. For
items, all unIDed spellbooks are lightmagenta, unIDed manuals lightcyan,
IDed spellbooks are magenta, and fully IDed manuals are cyan. In-menu,
IDed manuals are cyan with spellbooks simply in white. All book subtypes
are now visually distinguishable, including based on identification
info.