RandomTiles: Fix state initialization and improve comments
The index state was being initialized to 0 when the state table is
first created, which isn't a valid tile index, leading to lua errors
in enable_random_tile() the first time a user loaded RandomTiles. Now
when initializing, we properly initialize the index state to a random
tile and record the other relevant state values.