RandomTiles: Updates for mutant beasts
Add the new set of base tiles for mutant beasts for 0.18+ as a fixed
choice from the variants, with "mutant fire beast" also available for
the set of fire variants. Sadly the full set of facets can't be used to
make a legitimate mutant beast, as this would require combining multiple
tiles, which tile_player_tile can't do.