RandomTiles: Fix issues with state, RC cooperation, and indent
Previously the tile timer wouldn't change the tile for entries with
"fixed" var_type, even if they were the only entry in player_tiles. Now
when the timer triggers, we choose a new variant for "fixed" entries
even if the tile index is unchanged.

For RC files we now don't overwrite player_tiles when it's already been
defined, which allows people defining their own tilesets to using
'include += RandomTiles.rc' without having to copy all the RandomTiles
lua into their RC. For this to work, it seems that you need to load
RandomTiles.rc with the include after defining player_tiles, since crawl
doesn't seem to be honoring the processing order of '<>', which
RandomTiles.rc now uses. I'll check into this later on the crawl side of
things.

Finally, reindent to use the two-space indent rule, since this style is
used almost everywhere else in crawl land. The other lua files in this
repo will get reindented eventually as well.