RandomTiles: Generalize var_type and add Bai Suzhen
Previously "status" and "terrain" var_type values had their own code
that performed a similar check against a different player_tile entry
structure. This commit generalizes the handling of those value and adds
some new ones: "percent_hp" and "xl". There is now a var_functions table
for all of these conditional var_type types that gives the function to
use in order to get the player's current value.

These var types now expect tileset entries with the same structure
previously used for "status" and will apply the tileset entry whose
condition matches the player value. For string-based var types like
"status" and "terrain", the first match is used to allow order or
preference. For numeric var types like "percent_hp" and "xl", the entry
matches if the player's value is less than or equal to the condition
value and greater than any other value. For all var types, the first
entry is the default value applied when no other entry matches.

I didn't populate var_functions entries for all the other fields
available in the "you" lua library, since they're not being used in any
default entries, but it might be nice to add these in the future.

This commit adds Bai Suzhen with a condition on the player's percent hp,
so that dragon tile is used at or below 50%. The other terrain-based
merfolk tiles are converted to the new entry format, and the "siren"
entry now uses terrain instead of being fixed.  The structure used by
"status" is generally better since it allows us to have a