gammafunk.rc: Updated colors for 0.20
With many potion and wand changes, I've reorganized the colors for 0.20.
The rc checks for 0.19, loading the old ones if necessary. Escape
potions/scrolls are now on green, magic potions on magenta, enchantment
also on magenta. I've generally removed use of lightred with the
exception for ultra-rare potions of experience. It's good to minimize
use of this color, since it tends to be confused with ordinary red, the
color used for things that cause penance.

Also enable console equip_bar.