Rework default force_mores in force_mores.lua and those in gammafunk.rc
Reduce the scope of what we force_more to a few classes of dangerous
ranged effects or fast things with very high damage. The ones we would
care about all game go to gammafunk.rc.

For the dynamic force_mores, restrict these to things encountered in the
dungeon that can do close to 50% of current max HP with a ranged attack.
Adders and hounds get a rule for MHP <= 30 due to their speed.