Move speedrun options/code to gammaspeed_base.rc and gammaspeed.rc
This has macros and lua useful for some kinds of speedruns. It includes
gammafunk.rc automatically as it depends on that RC.

This RC will be uploaded as the RC file of the account gammaspeed on
webtiles servers, just like gammafunk.rc and RandomTiles.rc, and will be
kept up to date as long as those are.