Remove target_skill and the other files accordingly
Now that skill training targets exist in 0.21, this code is obsolete.
Update char_defaults to work without target skill and remove
target_skill-related code from it. Have char_defaults send 'm' to open
the skill menu once on turn 0 if it couldn't find a char profile. Also
send 'm' once on turn 0 in my RC for users other than myself.

Replace the target_skill setting macro on ']' with the function to set
the char profile for the current combo with char_defaults. This could be
improved to automatically save these settings after the 'm' screen is
finished, similar to how target_skill behaved, but that might be tricky
since the 'm' screen is an actual menu and not a prompt.

I'll probably make a branch for 0.20 at some point and move this code
and my RC for that version there, for reference, since some people may
prefer it and it's a good code reference to have easilly accessible.