gammafunk.rc: Updates to spell slots
No longer put the starting spell on 'z', since this is both annoying to
manage for slots and not especially helpful while playing in practice.
Many updates to spell slots due to spell additionals and removals, as
well as new letters for many spells so that we're less likely to not
have any available.