feat: Generalize inline-rc.sh
Make this script more applicable to other projects. Remove the `-d` and
`-b` arguments, requiring the rcfile as the first argument and the lua
files to insert as subsequent arguments. Users can specify full paths
for the lua files or run the script from the lua directory with relative
paths, using an absolute path for the RC. The `-o` option has a default
of the input rc file with the '.new' extension added to it.

Allow the replacement marker to be specified with `-m`, using a default
of '# BEGIN LUA'. We move the required RC lua block symbols `{` and `}`
into the base RC file itself so that users of this script can opt to put
those symbols in their lua files if they wish.